+++ index.html
@@ -24,6 +24,6 @@
             </div>
         </div>
     </div>
-    <script src="toDo.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
